+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-
 ## v0.4.2 Beta 1 (WIP)
-- [Compile] Make rebuild_if_exist default to False
 - 去除对opencv的依赖（WIP）
 - openvino backend （WIP）
+
+## v0.4.2 Beta 1 
+- [Compile] Make rebuild_if_exist default to False
 - Mat2Tensor support data_format now
 
 ## v0.4.1 Beta 2 (2024.1.17)